% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_non_zero_coeff}
\alias{determine_non_zero_coeff}
\title{Determining non-zero elastic net coefficients}
\usage{
determine_non_zero_coeff(enet_coefficients)
}
\arguments{
\item{enet_coefficients}{coefficients from a fitted elastic net model}
}
\value{
A vector of the best features
}
\description{
Determining non-zero elastic net coefficients
}
\examples{
enet_coefficients <- c(0,0,0.1,0.2,0,0)
determine_non_zero_coeff(enet_coefficients)
}
