% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extractorRData}
\alias{extractorRData}
\title{Function for extracting an object from a .RData file created by R's save() command
Inputs: RData file, object name}
\usage{
extractorRData(file, object)
}
\description{
Function for extracting an object from a .RData file created by R's save() command
Inputs: RData file, object name
}
