% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.r
\name{get_traits}
\alias{get_traits}
\title{Get traits based on the specified trait_id.}
\arguments{
\item{data}{The input data containing traits.}

\item{trait_id}{The identifier for the trait to load.}

\item{proj_dir}{The project directory path.}

\item{n_sub}{The number of subjects.}
}
\value{
A vector of traits.
}
\description{
Get traits based on the specified trait_id.
}
