% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity.R
\name{print_info}
\alias{print_info}
\title{Prediction information printing}
\usage{
print_info(trait, n_feat, n_sub, perc_tn, prof, tap, cov, run_svc)
}
\arguments{
\item{trait}{Target trait to predict}

\item{n_feat}{Number of features to use to predict}

\item{n_sub}{Number of subjects}

\item{perc_tn}{Percentage of training subjects}

\item{prof}{Use profile likelihood}

\item{tap}{Use covariance tapering}

\item{cov}{Which covariance function to use}

\item{run_svc}{Whether to run varying-coefficient model}
}
\value{
Print information
}
\description{
Prediction information printing
}
\examples{
trait <- 1; n_feat <- 1436; n_sub <- 5000; perc_tn <- 90; prof <- 1;
tap <- 5; cov = "exp"; run_svc <- 1
print_info(trait, n_feat, n_sub, perc_tn, prof, tap, cov, run_svc)
}
