% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.r
\name{run_elastic_net_model}
\alias{run_elastic_net_model}
\title{Run Elastic Net Model}
\usage{
run_elastic_net_model(idps_linear, trait, id_train, age, model_age)
}
\arguments{
\item{idps_linear}{A matrix of IDP (input data) features.}

\item{trait}{A vector of target variable (traits) to predict.}

\item{id_train}{Indices of training data.}

\item{age}{A vector of ages corresponding to the observations.}

\item{model_age}{Integer to specify how age is treated in the model:
0 = No age, 1 = Age as a covariate, 2 = Age only, 3 = IDPs + Age.}
}
\value{
A list containing:
\item{cvfit_glmnet}{The fitted Elastic Net model.}
\item{yhat_train}{Predictions for the training data.}
\item{yhat_test}{Predictions for the test data.}
}
\description{
Fits an Elastic Net model with or without age as a covariate and returns the fitted model
and predicted values for both training and test sets.
}
