% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.r
\name{run_lgboost_model_cv}
\alias{run_lgboost_model_cv}
\title{Run LightGBM with Cross-Validation}
\usage{
run_lgboost_model_cv(df_boost, params = NULL, n_folds = 10)
}
\arguments{
\item{df_boost}{A data frame containing features and target variable ("y").}

\item{params}{A list of LightGBM parameters (optional).}

\item{n_folds}{Number of cross-validation folds.}
}
\value{
A list containing:
\item{predictions_test}{Predicted values for all samples across the test folds.}
\item{corr_train}{Training set correlations for each fold.}
\item{corr_test}{Test set correlations for each fold.}
\item{mse_train}{Mean squared error on the training set for each fold.}
\item{mse_test}{Mean squared error on the test set for each fold.}
\item{r_squared_train}{R-squared values for the training set for each fold.}
\item{r_squared_test}{R-squared values for the test set for each fold.}
}
\description{
Performs k-fold cross-validation using LightGBM, fitting the model on the training data
and evaluating it on the test data.
}
