% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.r
\name{run_linear_model}
\alias{run_linear_model}
\title{Run Linear Model on Train and Test Data}
\usage{
run_linear_model(df_linear, id_train)
}
\arguments{
\item{df_linear}{A data frame containing the dependent variable \code{y} and the independent variables (features).}

\item{id_train}{A numeric vector indicating the indices of the training data within the data frame.}
}
\value{
A list containing:
\item{fit_lm}{The fitted linear model object.}
\item{lm_yhat}{A numeric vector of predictions for both training and test data.}
}
\description{
This function fits a linear model on the training data and uses it to predict
outcomes for both the training and test data. The predictions are returned alongside
the fitted linear model object.
}
\examples{
result <- run_linear_model(df_linear, id_train)
print(result$fit_lm)
}
