% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.r
\name{run_xgboost_model}
\alias{run_xgboost_model}
\title{Run XGBoost Model}
\usage{
run_xgboost_model(df_all_train_x, df_all_train_y, df_all_test_x, df_all_test_y)
}
\arguments{
\item{df_all_train_x}{A matrix or data frame containing the training data features.}

\item{df_all_train_y}{A vector containing the training data labels.}

\item{df_all_test_x}{A matrix or data frame containing the test data features.}

\item{df_all_test_y}{A vector containing the test data labels.}
}
\value{
A list containing:
\itemize{
\item predictions_train: Predicted values for the training data.
\item predictions_test: Predicted values for the test data.
\item xgb_model: The fitted XGBoost model object.
}
}
\description{
Trains an XGBoost model using the provided training data and tests the model on the provided test data.
}
\examples{
result <- run_xgboost_model(train_x, train_y, test_x, test_y)
predictions <- result$predictions_test
}
